Core and ORM are separated.
