+++ hep_analysis/settings.py
@@ -11,3 +11,4 @@ BASE_DIR = Path(__file__).resolve().parent.parent
 # connection string to sqlite database
 DB_PATH =  os.path.join(BASE_DIR, 'db.sqlite3')
 DB_CONNECTION_STRING  = "sqlite:///" + DB_PATH
+
